fix: extract artifacts in reconciler and deduplicate parsing logic
- Create tekton_utils.go with shared ParsePlatform(), ParseGitHash(), ConvertOciArtifacts(), ParseTektonImage()
- Update cloudevent.go to use exported functions from rest/service
- Update pipelineRunToTektonPipeline() to extract OciArtifacts and Images
- Update tests to verify artifact extraction

Ref: FLA-217 PR2 Architect review fixes
